web/front/requirements.txt
flask==1.1.1
werkzeug==0.15.4
python-dotenv==0.10.3